+++ std-bits/table/src/main/java/org/enso/table/data/column/operation/map/numeric/arithmetic/MinOp.java
@@ -30,7 +30,7 @@ public class MinOp<T extends Number, I extends Storage<? super T>>
 
   @Override
   public BigDecimal doBigDecimal(
-      BigDecimal a, BigDecimal b, int ix, MapOperationProblemAggregator problemAggregator) {
+      BigDecimal a, BigDecimal b, long ix, MapOperationProblemAggregator problemAggregator) {
     return a.min(b);
   }
 }
