+++ std-bits/table/src/main/java/org/enso/table/data/column/operation/map/numeric/arithmetic/NumericBinaryOpDefinition.java
@@ -13,5 +13,5 @@ public interface NumericBinaryOpDefinition {
       BigInteger a, BigInteger b, int ix, MapOperationProblemAggregator problemAggregator);
 
   BigDecimal doBigDecimal(
-      BigDecimal a, BigDecimal b, int ix, MapOperationProblemAggregator problemAggregator);
+      BigDecimal a, BigDecimal b, long ix, MapOperationProblemAggregator problemAggregator);
 }
